+++ src/modules/ebooks/ebooks.controller.ts
@@ -9,6 +9,8 @@ import {
   HttpCode,
   HttpStatus,
   UseGuards,
+  UploadedFiles,
+  UseInterceptors,
 } from '@nestjs/common';
 import { EbooksService } from './ebooks.service';
 import { CreateEbookDto } from './dto/create-ebook.dto';
@@ -23,11 +25,17 @@ import { Ebook } from './entities/ebook.entity';
 import { Roles } from '../roles/roles.decorator';
 import { AuthGuard } from '@nestjs/passport';
 import { RolesGuard } from 'src/common/guards/roles.guard';
+import { FileFieldsInterceptor } from '@nestjs/platform-express';
+import { SupabaseService } from 'src/supabase';
+import { ContentFileEnum } from '../contents/content-file.enum';
 
 @ApiTags('Contents')
 @Controller({ path: 'api/contents/ebooks', version: '1' })
 export class EbooksController {
-  constructor(private readonly ebooksService: EbooksService) {}
+  constructor(
+    private readonly ebooksService: EbooksService,
+    private readonly supabaseService: SupabaseService,
+  ) {}
 
   @Post()
   @UseGuards(AuthGuard('jwt'), RolesGuard)
@@ -40,9 +48,77 @@ export class EbooksController {
     description: 'The record has been successfully created.',
   })
   @ApiResponse({ status: 403, description: 'Forbidden.' })
+  @UseInterceptors(
+    FileFieldsInterceptor([{ name: 'thumbnail' }, { name: 'file_url' }]),
+  )
   @HttpCode(HttpStatus.CREATED)
-  create(@Body() createEbookDto: CreateEbookDto) {
-    return this.ebooksService.create(createEbookDto);
+  async create(
+    @UploadedFiles()
+    files: {
+      thumbnail?: Express.Multer.File[];
+      file_url?: Express.Multer.File[];
+    },
+    @Body() createEbookDto: CreateEbookDto,
+  ) {
+    let thumbFilename: string;
+    let fileFilename: string;
+    try {
+      if (files.thumbnail && files.thumbnail.length > 0) {
+        const {
+          success: thumbnailSuccess,
+          url: thumbnailUrl,
+          fileName: thumbnailFilename,
+          error: thumbnailError,
+        } = await this.supabaseService.uploadFile(
+          files.thumbnail[0],
+          `skilins_storage/${ContentFileEnum.thumbnail}`,
+        );
+
+        if (!thumbnailSuccess) {
+          throw new Error(`Failed to upload thumbnail: ${thumbnailError}`);
+        }
+
+        thumbFilename = thumbnailFilename;
+
+        createEbookDto.thumbnail = thumbnailUrl;
+      }
+
+      if (files.file_url && files.file_url.length > 0) {
+        const {
+          success: fileSuccess,
+          url: fileUrl,
+          fileName: fileUrlFilename,
+          error: fileError,
+        } = await this.supabaseService.uploadFile(
+          files.file_url[0],
+          `skilins_storage/${ContentFileEnum.file_ebook}`,
+        );
+
+        if (!fileSuccess) {
+          throw new Error(`Failed to upload file: ${fileError}`);
+        }
+        fileFilename = fileUrlFilename;
+        createEbookDto.file_url = fileUrl;
+      }
+      const result = await this.ebooksService.create(createEbookDto);
+      return result;
+    } catch (e) {
+      console.error('Error during ebook podcast creation:', e.message);
+
+      const { success, error } = await this.supabaseService.deleteFile([
+        `${ContentFileEnum.file_ebook}${fileFilename}`,
+        `${ContentFileEnum.thumbnail}${thumbFilename}`,
+      ]);
+
+      if (!success) {
+        console.error('Failed to delete files:', error);
+      }
+
+      return {
+        message:
+          'Failed to create ebook podcast and cleaned up uploaded files.',
+      };
+    }
   }
 
   @Get()
@@ -70,9 +146,51 @@ export class EbooksController {
   @ApiOkResponse({
     type: Ebook,
   })
+  @UseInterceptors(
+    FileFieldsInterceptor([{ name: 'thumbnail' }, { name: 'file_url' }]),
+  )
   @HttpCode(HttpStatus.OK)
-  update(@Param('uuid') uuid: string, @Body() updateEbookDto: UpdateEbookDto) {
-    return this.ebooksService.update(uuid, updateEbookDto);
+  async update(
+    @Param('uuid') uuid: string,
+    files: {
+      thumbnail?: Express.Multer.File[];
+      file_url?: Express.Multer.File[];
+    },
+    @Body() updateEbookDto: UpdateEbookDto,
+  ) {
+    const ebook = await this.ebooksService.update(uuid, updateEbookDto);
+
+    if (ebook.status === 'success') {
+      const isExist = await this.ebooksService.findOne(uuid);
+
+      if (files.thumbnail && files.thumbnail.length > 0) {
+        const thumbFilename = isExist.data.thumbnail.split('/').pop();
+
+        const { success: thumbnailSuccess, error: thumbnailError } =
+          await this.supabaseService.updateFile(
+            `${ContentFileEnum.thumbnail}${thumbFilename}`,
+            files.thumbnail[0],
+          );
+
+        if (!thumbnailSuccess) {
+          throw new Error(`Failed to update thumbnail: ${thumbnailError}`);
+        }
+      }
+
+      if (files.file_url && files.file_url.length > 0) {
+        const fileFilename = isExist.data.file_url.split('/').pop();
+        const { success: fileSuccess, error: fileError } =
+          await this.supabaseService.updateFile(
+            `${ContentFileEnum.file_ebook}${fileFilename}`,
+            files.file_url[0],
+          );
+
+        if (!fileSuccess) {
+          throw new Error(`Failed to update file: ${fileError}`);
+        }
+      }
+    }
+    return ebook;
   }
 
   @Delete(':uuid')
@@ -82,7 +200,30 @@ export class EbooksController {
     type: Ebook,
   })
   @HttpCode(HttpStatus.OK)
-  remove(@Param('uuid') uuid: string) {
-    return this.ebooksService.remove(uuid);
+  async remove(@Param('uuid') uuid: string) {
+    const isExist = await this.ebooksService.findOne(uuid);
+    const thumbFilename = isExist.data.thumbnail.split('/').pop();
+    const fileFilename = isExist.data.file_url.split('/').pop();
+    if (isExist) {
+      const ebook = await this.ebooksService.remove(uuid);
+      if (ebook.status === 'success') {
+        const { success: successThumb, error: errorThumb } =
+          await this.supabaseService.deleteFile([
+            `${ContentFileEnum.thumbnail}${thumbFilename}`,
+          ]);
+
+        if (!successThumb) {
+          console.error('Failed to delete thumbnail:', errorThumb);
+        }
+        const { success, error } = await this.supabaseService.deleteFile([
+          `${ContentFileEnum.file_ebook}${fileFilename}`,
+        ]);
+
+        if (!success) {
+          console.error('Failed to delete file:', error);
+        }
+      }
+      return ebook;
+    }
   }
 }
