+++ src/modules/pkl-reports/pkl-reports.controller.ts
@@ -9,6 +9,8 @@ import {
   HttpCode,
   HttpStatus,
   UseGuards,
+  UseInterceptors,
+  UploadedFiles,
 } from '@nestjs/common';
 import { PklReportsService } from './pkl-reports.service';
 import { CreatePklReportDto } from './dto/create-pkl-report.dto';
@@ -18,11 +20,17 @@ import { PklReport } from './entities/pkl-report.entity';
 import { AuthGuard } from '@nestjs/passport';
 import { RolesGuard } from 'src/common/guards/roles.guard';
 import { Roles } from '../roles/roles.decorator';
+import { FileFieldsInterceptor } from '@nestjs/platform-express';
+import { ContentFileEnum } from '../contents/content-file.enum';
+import { SupabaseService } from 'src/supabase';
 
 @ApiTags('Contents')
 @Controller({ path: 'api/contents/reports', version: '1' })
 export class PklReportsController {
-  constructor(private readonly pklReportsService: PklReportsService) {}
+  constructor(
+    private readonly pklReportsService: PklReportsService,
+    private readonly supabaseService: SupabaseService,
+  ) {}
 
   @Post()
   @UseGuards(AuthGuard('jwt'), RolesGuard)
@@ -30,9 +38,77 @@ export class PklReportsController {
   @ApiCreatedResponse({
     type: PklReport,
   })
+  @UseInterceptors(
+    FileFieldsInterceptor([{ name: 'thumbnail' }, { name: 'file_url' }]),
+  )
   @HttpCode(HttpStatus.CREATED)
-  create(@Body() createPklReportDto: CreatePklReportDto) {
-    return this.pklReportsService.create(createPklReportDto);
+  async create(
+    @UploadedFiles()
+    files: {
+      thumbnail?: Express.Multer.File[];
+      file_url?: Express.Multer.File[];
+    },
+    @Body() createPklReportDto: CreatePklReportDto,
+  ) {
+    let thumbFilename: string;
+    let fileFilename: string;
+    try {
+      if (files.thumbnail && files.thumbnail.length > 0) {
+        const {
+          success: thumbnailSuccess,
+          url: thumbnailUrl,
+          fileName: thumbnailFilename,
+          error: thumbnailError,
+        } = await this.supabaseService.uploadFile(
+          files.thumbnail[0],
+          `skilins_storage/${ContentFileEnum.thumbnail}`,
+        );
+
+        if (!thumbnailSuccess) {
+          throw new Error(`Failed to upload thumbnail: ${thumbnailError}`);
+        }
+
+        thumbFilename = thumbnailFilename;
+
+        createPklReportDto.thumbnail = thumbnailUrl;
+      }
+
+      if (files.file_url && files.file_url.length > 0) {
+        const {
+          success: fileSuccess,
+          url: fileUrl,
+          fileName: fileUrlFilename,
+          error: fileError,
+        } = await this.supabaseService.uploadFile(
+          files.file_url[0],
+          `skilins_storage/${ContentFileEnum.file_report}`,
+        );
+
+        if (!fileSuccess) {
+          throw new Error(`Failed to upload file: ${fileError}`);
+        }
+        fileFilename = fileUrlFilename;
+        createPklReportDto.file_url = fileUrl;
+      }
+
+      const result = await this.pklReportsService.create(createPklReportDto);
+      return result;
+    } catch (e) {
+      console.error('Error during report podcast creation:', e.message);
+
+      const { success, error } = await this.supabaseService.deleteFile([
+        `${ContentFileEnum.file_report}${fileFilename}`,
+        `${ContentFileEnum.thumbnail}${thumbFilename}`,
+      ]);
+
+      if (!success) {
+        console.error('Failed to delete files:', error);
+      }
+
+      return {
+        message: 'Failed to create report and cleaned up uploaded files.',
+      };
+    }
   }
 
   @Get()
@@ -57,15 +133,58 @@ export class PklReportsController {
   @Patch(':uuid')
   @UseGuards(AuthGuard('jwt'), RolesGuard)
   @Roles('admin')
+  @UseInterceptors(
+    FileFieldsInterceptor([{ name: 'thumbnail' }, { name: 'file_url' }]),
+  )
   @ApiOkResponse({
     type: PklReport,
   })
   @HttpCode(HttpStatus.OK)
-  update(
+  async update(
     @Param('uuid') uuid: string,
+    @UploadedFiles()
+    files: {
+      thumbnail?: Express.Multer.File[];
+      file_url?: Express.Multer.File[];
+    },
     @Body() updatePklReportDto: UpdatePklReportDto,
   ) {
-    return this.pklReportsService.update(uuid, updatePklReportDto);
+    const report = await this.pklReportsService.update(
+      uuid,
+      updatePklReportDto,
+    );
+    if (report.status === 'success') {
+      const isExist = await this.pklReportsService.findOne(uuid);
+
+      if (files.thumbnail && files.thumbnail.length > 0) {
+        const thumbFilename = isExist.data.thumbnail.split('/').pop();
+
+        const { success: thumbnailSuccess, error: thumbnailError } =
+          await this.supabaseService.updateFile(
+            `${ContentFileEnum.thumbnail}${thumbFilename}`,
+            files.thumbnail[0],
+          );
+
+        if (!thumbnailSuccess) {
+          throw new Error(`Failed to update thumbnail: ${thumbnailError}`);
+        }
+      }
+
+      if (files.file_url && files.file_url.length > 0) {
+        const fileFilename = isExist.data.file_url.split('/').pop();
+        const { success: fileSuccess, error: fileError } =
+          await this.supabaseService.updateFile(
+            `${ContentFileEnum.file_report}${fileFilename}`,
+            files.file_url[0],
+          );
+
+        if (!fileSuccess) {
+          throw new Error(`Failed to update file: ${fileError}`);
+        }
+      }
+    }
+
+    return report;
   }
 
   @Delete(':uuid')
@@ -75,7 +194,30 @@ export class PklReportsController {
     type: PklReport,
   })
   @HttpCode(HttpStatus.OK)
-  remove(@Param('uuid') uuid: string) {
-    return this.pklReportsService.remove(uuid);
+  async remove(@Param('uuid') uuid: string) {
+    const isExist = await this.pklReportsService.findOne(uuid);
+    const thumbFilename = isExist.data.thumbnail.split('/').pop();
+    const fileFilename = isExist.data.file_url.split('/').pop();
+    if (isExist) {
+      const report = await this.pklReportsService.remove(uuid);
+      if (report.status === 'success') {
+        const { success: successThumb, error: errorThumb } =
+          await this.supabaseService.deleteFile([
+            `${ContentFileEnum.thumbnail}${thumbFilename}`,
+          ]);
+
+        if (!successThumb) {
+          console.error('Failed to delete thumbnail:', errorThumb);
+        }
+        const { success, error } = await this.supabaseService.deleteFile([
+          `${ContentFileEnum.file_report}${fileFilename}`,
+        ]);
+
+        if (!success) {
+          console.error('Failed to delete file:', error);
+        }
+      }
+      return report;
+    }
   }
 }
